Projects page cards

+++ src/components/Contact/index.js
@@ -7,12 +7,23 @@ import './index.scss'
 import { ToastContainer, toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 import { useTranslation } from 'react-i18next'
+import ReCAPTCHA from 'react-google-recaptcha'
 
 const Contact = () => {
   const { t } = useTranslation()
   const [letterClass, setLetterClass] = useState('text-animate')
   const form = useRef()
   const contactMeArray = t('contact_me').split('')
+  const [confirmEmail, setConfirmEmail] = useState(false)
+  const [confirmCaptcha, setConfirmCaptcha] = useState(null)
+  const captchaSiteKey = '6LfMoGkhAAAAABzZu2RTPtU1JibMGIZRUI20ogQv'
+
+  const handleConfirmEmail = () => {
+    setConfirmEmail(!confirmEmail)
+  }
+  const handleCaptcha = (value) => {
+    setConfirmCaptcha(value)
+  }
 
   useEffect(() => {
     async function changeClass() {
@@ -26,33 +37,51 @@ const Contact = () => {
   useEffect(() => {
     const response = localStorage.getItem('res')
     if (response === 'success') {
-      toast.success("Your message has been sent!", {
-        theme: "colored",
-        icon: "✉️"
-      });
+      toast.success('Your message has been sent!', {
+        theme: 'colored',
+        icon: '✉️',
+      })
       localStorage.removeItem('res')
     }
   }, [])
 
   const sendEmail = (e) => {
     e.preventDefault()
-
-    emailjs
-      .sendForm(
-        'service_kyflqxk',
-        'template_vdsk90s',
-        form.current,
-        'gt3Zfth5Wt6lJOCcX'
-      )
-      .then(
-        () => {
-          localStorage.setItem('res', 'success')
-          window.location.reload(false)
-        },
-        () => {
-          alert('Something went wrong, please try again!')
-        }
-      )
+    if (confirmEmail === false) {
+      emailjs
+        .sendForm(
+          'service_kyflqxk',
+          'template_vdsk90s',
+          form.current,
+          'gt3Zfth5Wt6lJOCcX'
+        )
+        .then(
+          () => {
+            localStorage.setItem('res', 'success')
+            window.location.reload(false)
+          },
+          () => {
+            alert('Something went wrong, please try again!')
+          }
+        )
+    } else {
+      emailjs
+        .sendForm(
+          'service_kyflqxk',
+          'template_7odrvlr',
+          form.current,
+          'gt3Zfth5Wt6lJOCcX'
+        )
+        .then(
+          () => {
+            localStorage.setItem('res', 'success')
+            window.location.reload(false)
+          },
+          () => {
+            alert('Something went wrong, please try again!')
+          }
+        )
+    }
   }
 
   return (
@@ -67,11 +96,13 @@ const Contact = () => {
                 idx={15}
               />
             </h1>
-            <p>
-              {t('contact_text')}
-            </p>
+            <p>{t('contact_text')}</p>
             <div className="contact-form">
               <form ref={form} onSubmit={sendEmail}>
+                <div
+                  class="g-recaptcha"
+                  data-sitekey="6LfMoGkhAAAAABzZu2RTPtU1JibMGIZRUI20ogQv"
+                ></div>
                 <ul>
                   <li className="half">
                     <input
@@ -109,7 +140,32 @@ const Contact = () => {
                     ></textarea>
                   </li>
                   <li>
-                    <input type="submit" className="flat-button" value={t('contact_send')} />
+                    <div className="form-control">
+                      <label className="label cursor-pointer">
+                        <span className="label-text">
+                          {t('contactPage.confirm_email')}
+                        </span>
+                        <input
+                          type="checkbox"
+                          checked={confirmEmail}
+                          onChange={handleConfirmEmail}
+                          className="checkbox bg-white"
+                        />
+                      </label>
+                    </div>
+                  </li>
+                  <li>
+                    <ReCAPTCHA
+                      sitekey={captchaSiteKey}
+                      onChange={handleCaptcha}
+                    />
+                  </li>
+                  <li>
+                    <input
+                      type="submit"
+                      className="flat-button"
+                      value={t('contact_send')}
+                    />
                   </li>
                 </ul>
               </form>
@@ -129,7 +185,7 @@ const Contact = () => {
         pauseOnFocusLoss
         draggable
         pauseOnHover={false}
-        style={{fontSize:'1rem'}}
+        style={{ fontSize: '1rem' }}
       />
     </>
   )
